+++ src/app/app.module.ts
@@ -1,24 +1,40 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { Routes, RouterModule } from "@angular/router";
 
 import { AppComponent } from './app.component';
-import { CurrentPageComponent } from './current-page/current-page.component';
+import { HomeComponent } from './home/home.component';
+import { CurrentTemperatureComponent } from './current-temperature/current-temperature.component';
 import { FiveDayForecastComponent } from './five-day-forecast/five-day-forecast.component';
 import { NavbarComponent } from './navbar/navbar.component';
+import { LocationService } from './services/location.service';
+import { NotFoundComponent } from './not-found/not-found.component';
+
+// Routes Angular needs to know to access other "pages" using the navbar.
+const appRoutes: Routes = [
+  { path: '', component: HomeComponent },
+  { path: 'current', component: CurrentTemperatureComponent },
+  { path: 'forecast', component: FiveDayForecastComponent },
+  { path: 'not-found', component: NotFoundComponent },
+  { path: '**', redirectTo: 'not-found'}
+];
 
 @NgModule({
   declarations: [
     AppComponent,
-    CurrentPageComponent,
+    CurrentTemperatureComponent,
     FiveDayForecastComponent,
-    NavbarComponent
+    NavbarComponent,
+    HomeComponent,
+    NotFoundComponent,
   ],
   imports: [
     BrowserModule,
-    FormsModule
+    FormsModule,
+    RouterModule.forRoot(appRoutes)
   ],
-  providers: [],
+  providers: [ LocationService ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }