+++ src/environments/environment.prod.ts
@@ -1,3 +1,6 @@
 export const environment = {
-  production: true
+  production: true,
+  appID: 'bf4b4ba5c71e63e094c62cc9a10efa56',
+  baseURL: 'http://api.openweathermap.org/data/2.5/',
+  units: 'imperial'
 };