+++ src/environments/environment.qa.ts
@@ -1,0 +1,6 @@
+export const environment = {
+    production: false,
+    appID: 'bf4b4ba5c71e63e094c62cc9a10efa56',
+    baseURL: 'http://api.openweathermap.org/data/2.5/',
+    units: 'imperial'
+  };