+++ src/environments/environment.ts
@@ -3,14 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  appID: 'bf4b4ba5c71e63e094c62cc9a10efa56',
+  baseURL: 'http://api.openweathermap.org/data/2.5/',
+  units: 'imperial'
 };
-
-/*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.