+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
+## [0.0.2] 2019 - 04 - 14
+### Added
+- Environment variables
+- "page" routing
+- `home` component with challenge description
+- Created service to get user's current location
+- Created a rerouting component, `not-found` for 404 errors.
+
+### Changed
+- `current-page` component to `current-temperature`
+
 ## [0.0.1] 2019 - 04 - 13
 ### Added
 - Basic Angular app generated using Angular CLI
-- Three generated components: `navbar`, `current-page`, and `five-day-forecast`
+- Three generated components: `navbar`, `current-temperature`, and `five-day-forecast`
 - `CHANGELOG.md` to record project changes
 - Sun with cloud favicon
 - "Weather Forecaster" as the app's title
 
-[0.0.1]: 
+[0.0.1]:
+[0.0.2]: